renamed transmission to transmissions
