+++ examples/transmissions/mime_parts.js
@@ -5,13 +5,16 @@ var key = 'YOURAPIKEY'
   , client = new SparkPost(key);
 
 var trans = {
-  recipients: [{ address: { email: 'john.doe@example.com' } }],
-  rfc822: 'Content-Type: text/plain\nFrom: From Envelope <from@example.com>\nSubject: Example Email\n\nHello World',
   from: 'From Envelope <from@example.com>',
-  subject: 'Example Email for RFC-822 Content'
+  recipients: [{ address: { email: 'john.doe@example.com' } }],
+  subject: 'Example Email for MIME Parts',
+  html: '<html><body><p>Hello World!</p></body></html>',
+  text: 'Hello World!',
+  trackOpens: true,
+  trackClicks: true
 };
 
-client.transmission.send(trans, function(err, res) {
+client.transmissions.send(trans, function(err, res) {
   if (err) {
     console.log(err);
   } else {
