+++ examples/transmissions/rfc822.js
@@ -5,14 +5,13 @@ var key = 'YOURAPIKEY'
   , client = new SparkPost(key);
 
 var trans = {
+  recipients: [{ address: { email: 'john.doe@example.com' } }],
+  rfc822: 'Content-Type: text/plain\nFrom: From Envelope <from@example.com>\nSubject: Example Email\n\nHello World',
   from: 'From Envelope <from@example.com>',
-  subject: 'Example Email for Stored List and Template',
-  recipientList: 'example-list',
-  template: 'my-template',
-  recipients: [{ address: { email: 'john.doe@example.com' } }]
+  subject: 'Example Email for RFC-822 Content'
 };
 
-client.transmission.send(trans, function(err, res) {
+client.transmissions.send(trans, function(err, res) {
   if (err) {
     console.log(err);
   } else {
