+++ examples/transmissions/stored_recipients_inline_content.js
@@ -12,7 +12,7 @@ var trans = {
   text: 'Hello World!'
 };
 
-client.transmission.send(trans, function(err, res) {
+client.transmissions.send(trans, function(err, res) {
   if (err) {
     console.log(err);
   } else {
