+++ examples/transmissions/stored_template_send.js
@@ -11,7 +11,7 @@ var trans = {
   recipients: [{ address: { email: 'john.doe@example.com' } }]
 };
 
-client.transmission.send(trans, function(err, res) {
+client.transmissions.send(trans, function(err, res) {
   if (err) {
     console.log(err);
   } else {
