+++ lib/SendGridCompatibility/index.js
@@ -119,7 +119,7 @@ var translatePayload = function(payload) {
  */
 sendgrid.prototype.send = function(payload, callback) {
   var translated = translatePayload(payload);
-  this.client.transmission.send(translated, callback);
+  this.client.transmissions.send(translated, callback);
 };
 
 sendgrid.prototype.Email = require('./Email');
