+++ lib/sparkpost.js
@@ -59,7 +59,7 @@ var SparkPost = function(apiKey, options) {
   this.sendingDomains = require('./sendingDomains')(this);
   this.suppressionList = require('./suppressionList')(this);
   this.templates = require('./templates')(this);
-  this.transmission = require('./transmission')(this);
+  this.transmissions = require('./transmissions')(this);
   this.webhooks = require('./webhooks')(this);
 };
