+++ lib/transmissions.js
